+++ src/tools/utils.jl
@@ -45,12 +45,12 @@ end
 """ 
     print_time(EC::AbstractECInfo, t1, info::AbstractString, verb::Int)
 
-  Print time with message `info` if verbosity `verb` is smaller than EC.verbosity.
+  Print time with message `info` if verbosity `verb` is smaller than `PrintOptions.time`.
 """
 function print_time(EC::AbstractECInfo, t1, info::AbstractString, verb::Int)
   t2 = time_ns()
-  if verb < EC.verbosity
-    output_time(t2-t1, info)
+  if verb < EC.options.print.time
+    output_time(t2 - t1, info)
   end
   return t2
 end